% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_edge.R
\name{gen_edge}
\alias{gen_edge}
\title{Creates edge List}
\usage{
gen_edge(nbatches, batch_len, Ngenes, mat)
}
\arguments{
\item{nbatches}{Number of batches to consider.}

\item{batch_len}{size of each batches}

\item{Ngenes}{Number of genes to consider for each batch}

\item{mat}{The preprocessed matrix with geans in each column and samples in every row.}
}
\value{
Returns the edgelist
}
\description{
Creates edge List
}
