% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphcluster.R
\docType{package}
\name{graphcluster}
\alias{graphcluster}
\title{Batchwise Clustering using Graphs Package}
\description{
Contains the functions that are deployed to perform the final clustering.
}
\author{
Riddhi Ghosal, Debajyoti Sinha \email{jojodigi@gmail.com}
}
