% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphcommunity.R
\name{graphcommunity}
\alias{graphcommunity}
\title{Form the final Graph communities}
\usage{
graphcommunity(edgelist, data_labels)
}
\arguments{
\item{edgelist}{The list of edges in the subgraphs}

\item{data_labels}{The data labels exported}
}
\description{
Form the final Graph communities
}
