% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_cluster.R
\name{parallel_cluster}
\alias{parallel_cluster}
\title{Creates edge List in Parallel}
\usage{
parallel_cluster(nbatches, batch_len, Ngenes, mat, prob = 0.65)
}
\arguments{
\item{nbatches}{Number of batches to consider.}

\item{batch_len}{size of each batches}

\item{Ngenes}{Number of genes to consider for each batch}

\item{mat}{The preprocessed matrix with geans in each column and samples in every row.}

\item{prob}{The quantile below which edges are not formed}
}
\value{
Returns the edgelist
}
\description{
Creates edge List in Parallel
}
