% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateweights_graph.R
\name{updateweights_graph}
\alias{updateweights_graph}
\title{Detects Clusters}
\usage{
updateweights_graph(edgelist, data_labels, prob = 0.6)
}
\arguments{
\item{edgelist}{Edgelist as returned by gen_edge function}

\item{data_labels}{The data labels exported}

\item{prob}{The quantile below which edges are not formed}
}
\value{
Returns graph density
}
\description{
Detects Clusters
}
